requirements.txt
Django>=4.2
djangorestframework
unidecode
fuzzywuzzy[speedup]
fasttext
requests
